refactor: Phase 4c sweep — dual-derive engine_types/common_types/transaction/storage + test fixes
Migrate the bulk of the host-side workspace to dual-derive
serde + minicbor::{Encode,Decode,CborLen} alongside existing serde
derives, with #[n(N)] tags in declaration order:

  - tari_engine_types: Substate, SubstateValue, SubstateId, Component,
    Resource, Vault, NonFungibleContainer, NonFungible,
    ClaimedOutputTombstone, TransactionReceipt, PublishedTemplate,
    ValidatorFeePool, Utxo, plus commit_result, events, fees,
    indexed_value, instruction_result and friends.
  - tari_ootle_common_types: ExtraData, ShardStateVersions (hand-rolled
    over BoundedVec), Shard, ShardGroup, MaxSizeBytes, plus state_version
    and versioned_substate_id.
  - tari_ootle_transaction: Transaction(V1), UnsealedTransaction,
    UnsignedTransaction, Instruction, Blobs, Assertion, NamedArg path
    bounds, plus the V1 sub-types.
  - tari_ootle_storage: BlockPledge, CommandsCommitProof, Evidence,
    EpochCheckpoint, Command and friends. Foreign tari_sidechain types
    ride along via #[cbor(with = "tari_bor::adapters::serde_bridge")].
  - tari_ootle_template_metadata: TemplateMetadata with option_url /
    option_oid adapter modules for Url and gix_hash::ObjectId.

tari_bor:
  - New adapters module: boxed_slice, indexset_codec, indexmap_codec, and
    serde_bridge (gated on serde, pulls in minicbor-serde so foreign
    serde-only types can be encoded inline via #[cbor(with = ...)]).
  - Value::is_unit() (matches Null OR empty Array) — minicbor encodes ()
    as an empty array, so callers checking "no value" need both forms.
  - Value::decoded::<T>() replaces the old ciborium-era .deserialized().
  - ByteCounter now also impls std::io::Write so host fee_module can
    measure encoded length through the same writer path.

Engine + integrations:
  - Auth-hook return check and Assertion::IsNotNull use is_unit().
  - engine_args, wasm process, instruction_result.decode, NonFungible
    decode helpers all switch from serde::DeserializeOwned to
    tari_bor::Decode<'_, ()>.
  - indexed_value::get_value_by_path no longer panics on non-numeric
    segments — returns None so legacy string-keyed paths fail cleanly.
  - p2p::encoding bounds flipped to minicbor; storage_sqlite drops the
    removed cbor_value_encoding_fix wrapper (now redundant).
  - transaction_manifest rejects 128-bit integer literals with a clear
    error (minicbor 2.2 has no Encode/Decode for u128/i128).

Engine tests:
  - Path-based extractions switched to integer indices ($.0/$.1/...) in
    access_rules, address_allocation, burn, freeze, recall, no_std and
    template_upgrade.
  - basic_nft, tickets, emoji_id, nft_list, template_upgrade/v2 and
    no_std test templates moved off serde-only derives.
  - template_test_tooling::builtin_component_state seeds the XTR / NFT
    faucet components as cbor!([...]) arrays to match the new struct
    encoding (single-field structs are 1-element arrays, not maps).
  - buggy template's hardcoded _ABI_TEMPLATE_DEF bytes regenerated for
    the new minicbor wire format.

+++ crates/engine/tests/templates/no_std/src/lib.rs
@@ -31,7 +31,10 @@ mod template {
 
     #[derive(Debug, Default)]
     pub struct NoStdCounter {
-        value: u128,
+        // u64 rather than u128: minicbor 2.2 does not impl Encode/Decode for u128/i128 and the
+        // workspace's bignum bridge (`Value::Integer(i128)` via `serde_bridge`) doesn't apply to
+        // bare template struct fields. The test exists to exercise no_std + allocator, not 128-bit.
+        value: u64,
     }
 
     impl NoStdCounter {
@@ -57,7 +60,7 @@ mod template {
             self.value += 1;
         }
 
-        pub fn reset_to(&mut self, value: u128) {
+        pub fn reset_to(&mut self, value: u64) {
             debug!("Changing value from {:?} to {}", self.value, value);
             self.value = value;
         }
